Moving static functions of main loop to a distinct source file. This is the first step to make amora a library (the idea is to have the CLI amora server using it and making easier to use it in an applet).
It required a change in the uloop code, to provide a context for registered
callbacks for each file descriptor. For while, there is only one common
context used for all callbacks (a logical improvement would be to have
a distinct context for each callback).

Also, the static functions were accessing a global instance of amora_s
structure. Now, this is a calling parameter.

There are several warnings and 1 utest is failing, but the server is already
functional with this change.