Makefile to build both, amora-cli and amora-applet in one go.
The Makefile is a sort of ugly hack since it uses sed to add some
library and include paths to amora-applet's Makefile after its
generation with qmake. But otherwise we would need to run "make
install" in amora-cli before -- which would not be nice for packaging,
neither because of the needed root rights nor because of having just
build libraries installed to continue.

In addition to some build targets, this make file also supports the
common targets all, clean, dist and distclean. The all target is the
default target and builds both, amora-applet and amora-cli.